moved from file(GLOB); added CMAKE_CURRENT_LIST_DIR to support include syntax
