+++ osquery/main/main.cpp
@@ -31,7 +31,7 @@
 #include "osquery/core/utils.h"
 #include "osquery/core/watcher.h"
 #include "osquery/devtools/devtools.h"
-#include "osquery/dispatcher/distributed.h"
+#include "osquery/dispatcher/distributed_runner.h"
 #include "osquery/dispatcher/io_service.h"
 #include "osquery/dispatcher/scheduler.h"
 #include "osquery/filesystem/fileops.h"
